* Attempt at a quick fix for compilation on 64 bit machines.
   I don't have a 64 bit machine, but I have received a report that
   there is a problem with the line I in this commit fix.

   There is a ambiguous overload in Dbg. My guess is that
   "int" and "long long" is the same size on that machine, but I'm not
   sure.



git-svn-id: svn://svn.minisip.org/minisip/trunk@3549 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/SipLayerDialog.cxx
@@ -67,7 +67,7 @@ bool SipLayerDialog::removeDialog(string callId){
 	size_t n = dialogs.erase(callId);
 #ifdef DEBUG_OUTPUT
 	if (n!=1){
-		merr << "WARNING: dialogs.erase should return 1, but returned "<< n<<endl;
+		merr << "WARNING: dialogs.erase should return 1, but returned "<< (short)n<<endl;
 	}
 #endif
 	return n==1;
